modules/article-harvest/src/article_harvest/time_utils.py
from __future__ import annotations

from datetime import date, datetime, timezone

from dateutil import parser


def iso_now() -> str:
    return datetime.utcnow().replace(microsecond=0).isoformat() + "Z"


def iso_date_today() -> str:
    return datetime.utcnow().date().isoformat()


def parse_date(value: str) -> date:
    try:
        return date.fromisoformat(value)
    except ValueError:
        dt = parser.parse(value)
        return dt.date()


def parse_datetime(value: str) -> datetime:
    try:
        dt = datetime.fromisoformat(value.replace("Z", "+00:00"))
    except ValueError:
        dt = parser.parse(value)
    if dt.tzinfo is None:
        return dt.replace(tzinfo=timezone.utc)
    return dt
