modules/article-harvest/pyproject.toml
[project]
name = "article-harvest"
version = "0.1.0"
description = "Local-first article harvest module for newsletter sources"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32",
  "beautifulsoup4>=4.12",
  "markdownify>=0.12",
  "lxml>=5.0",
  "python-dateutil>=2.9",
  "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
]

[project.scripts]
article-harvest = "article_harvest.cli:main"

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "C90"]

[tool.ruff.lint.mccabe]
max-complexity = 20
